net: fec: fix the potential memory leak in fec_enet_init()
stable inclusion
from stable-5.10.42
commit 8ee7ef4a57a9e1228b6f345aaa70aa8951c7e9cd
bugzilla: https://gitee.com/src-openeuler/kernel/issues/I9BGVY
CVE: CVE-2021-47150

----------------------------------------------

If the memory allocated for cbd_base is failed, it should
free the memory allocated for the queues, otherwise it causes
memory leak.

And if the memory allocated for the queues is failed, it can
return error directly.

Fixes: 59d0f7465644 ("net: fec: init multi queue date structure")
Signed-off-by: Fugang Duan <fugang.duan@nxp.com>
Signed-off-by: Joakim Zhang <qiangqing.zhang@nxp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Zhang Zekun <zhangzekun11@huawei.com>