arm64: Kconfig: select ARCH_SUSPEND_NONZERO_CPU by default
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I9NR7Q
CVE: NA

--------------------------------

Commit 08ae95f4fd3b ("nohz_full: Allow the boot CPU to be nohz_full")
allow CPU0 to be housekeeping CPU. But it is only supproted when
CONFIG_PM_SLEEP_SMP is not configured, or when it is configured and
the arch allows suspend on non-zero CPUs.

We cannot disable CONFIG_PM_SLEEP_SMP because freeze_secondary_cpus(),
which is used in kexec, is included in the config. For arm64, there is
not restriction in kernel to prevent nonzero cpu entering suspend using
PSCI interface. when ARCH_SUSPEND_NONZERO_CPU is enabled,
PM_SLEEP_SMP_NONZERO_CPU will be also enabled. The only influence is that
suspend_disable_secondary_cpus() will use a housekeeping CPU to enter
suspend. If pm sleep/hibernate is not involved, only kexec will use this
function, which is tested on a arm64 server, and the vmcore can be created
successfully. We add 'enable_cpu0_nohz_full' in the
CONFIG_PM_SLEEP_SMP_NONZERO_CPU check branch. When 'enable_cpu0_nohz_full'
is zero. There is no functional change.

Signed-off-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>