sched/numa: Fix numa imbalance in load_balance()
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I9RMHW
CVE: NA

--------------------------------

When perform load balance, a NUMA imbalance is allowed
if busy CPUs is less than the maximum threshold, it
keeps a pair of communication tasks on the current
node when the destination is lightly loaded.

1. But, calculate_imbalance() use local->sum_nr_running,
it may not be accurate, because communication tasks is
on busiest group, so should be busiest->sum_nr_running.

2. At the same time, idles cpus are used to calculate
imbalance, but the group_weight may not be the same
between local and busiest groups. In this case, even if
both groups are very idle, imbalance will be calculated
very large, so the imbalance is calculated by calculating
the difference of busy cpus between groups.

Signed-off-by: Zhang Qiao <zhangqiao22@huawei.com>