mm/memcontrol: add ksm state for memcg
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I9GT87

----------------------------------------

Add KSM state for memcg, the valid values include 0 and 1.

When changing auto_ksm_enabled from 0 to 1, enable KSM for tasks in the
memcg. When changing auto_ksm_enabled from 1 to 0, disable KSM for tasks
in the memcg. If enable/disable fails, return the error code and don't
change auto_ksm_enabled. If the auto_ksm_state of the child memcgs differ,
also enable/disable KSM for the tasks in the memcgs. If enable/disable
for a child memcg fails, stop traversing child memcgs and return the
error code.

When writing the value same to auto_ksm_enabled of the memcg, i.e. from 0
to 0 and 1 to 1, do nothing.

Signed-off-by: Jinjiang Tu <tujinjiang@huawei.com>