mm/swap: support to allocate page from hpool for swapcache
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I9H66Y
CVE: NA

--------------------------------

When swapin in async mode, if the swap entry is belong to a memcg and
the memcg is bound with a hpool, then we should try to allocate page
from hpool firstly. If the swapin is triggered in page fault, then we
don't need to allocate from hpool proactively, since the task is belong
to the memcg, the page will allocate from hpool naturally.

Signed-off-by: Liu Shixin <liushixin2@huawei.com>