RDMA/hns: Fix cpu stuck by printings during reset
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I9SI44

----------------------------------------------------------------------
During the reset, issuing the cmdq command to destroy resources
such as qp, cq, and mr may fail, and the destruction failure
information is printed. When a large number of resources fail to
be destroyed, printk() causes multiple threads to compete for a
global log buffer spinlock.  As a result, the CPU performance
deteriorates and the CPU is stuck. To fix it, use a function like
printk_limit_rate() to limit the print frequency of printk().

Fixes: 9a4435375cd1 ("IB/hns: Add driver files for hns RoCE driver")
Fixes: c7bcb13442e1 ("RDMA/hns: Add SRQ support for hip08 kernel mode")
Fixes: 626903e9355b ("RDMA/hns: Add support for reporting wc as software mode")
Fixes: 6eef524201de ("RDMA/hns: Replace not intuitive function/macro names")
Fixes: 6f5f556d3795 ("RDMA/hns: Use the reserved loopback QPs to free MR before destroying MPT")
Fixes: 357f34294686 ("RDMA/hns: Simplify the state judgment code of qp")
Fixes: 926a01dc000d ("RDMA/hns: Add QP operations support for hip08 SoC")
Fixes: 0273952c5e6e ("RDMA/hns: Add method to detach WQE buffer")
Fixes: 9c9edf689a60 ("RDMA/hns: Remove redundant DFX file and DFX ops structure")
Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: Guofeng Yue <yueguofeng@h-partners.com>