driver/arm/spe: making mem_sampling and perf mutually exclusive with spe driver
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I9GZAQ
CVE: NA

--------------------------------

The preorder patch constitutes a set of independent SPE drivers, and
isolates the basic operations of SPE from the perf driver, including
SPE interrupt handling, enabling and other operations, so that the use
of SPE functionality can be divided into two categories: the first is
the basic modules of the kernel, such as: Numa Balance, Damon, etc.,
and the second is the use of perf driver by the user, where the The
first type of users The first type of users starts and stops the SPE
through the abstraction layer mem_sampling. in addition, in order to
avoid as much as possible embedded modifications to the perf code, we
introduced the cmdline mem_sampling_on to control the use of
mem_sampling, when the perf spe can not be captured, and vice versa.

Signed-off-by: Ze Zuo <zuoze1@huawei.com>