crypto: jitter - change subsys_initcall(jent_mod_init) back to module_init(jent_mod_init)
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IA6GL4

--------------------------------

jent_entropy_init depend on the clocksource being initialized, with the
subsys_initcall(jent_mod_init) called in arm32, the initialization of
clocksource is after jent_mod_init, so change the subsys_initcall in
arm32 to module_init. And the fips depend on jitter init, this feature
is controlled by CONFIG_CRYPTO_FIPS, so use CONFIG_CRYPTO_FIPS to isolate
the case of the fips problem for arm64/x86 with the dependance of
clocksource for arm32.

Fixes: 0d9fbb7dcca5 ("crypto: jitter - change module_init(jent_mod_init) to subsys_initcall(jent_mod_init)")
Signed-off-by: Xiang Yang <xiangyang3@huawei.com>