virtcca feature: secure smmu init
virtcca inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IADD42

--------------------------------

VirtCCA Feature:
When the securecomponent, TMM (Trusted Management Monitor),
is loaded, the SMMU driver will initialize the secure SMMU
Every SMMU register setting will be forwarded to the secure
world for the corresponding security register settings.
Secure SMMU Init Process:
1. In the arm_smmu device probe function, it checks whether
TMM is loaded. If it is loaded, it will allocate an ID for
the SMMU that needs secure initialization. Otherwise, it will
not allocate an ID, and subsequent secure SMMU operations
will not be performed;
2. If TMM loaded, the ecmdq feature will not enable;
3. When initializing the queues in ARM SMMU, if the SMMU ID
is valid(TMM has been loaded), it will initialize the secure
memory corresponding to the queue in the secure world and set
the corresponding secure registers;
4. Finally, if the SMMU ID is valid, complete the enable
operation of the secure SMMU CRO register.

Signed-off-by: Xiangkai Yang <yangxiangkai@huawei.com>
---