Revert "kbuild: fix false positive warning/error about missing libelf"
openEuler inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IAD5GG?from=project-issue
CVE: NA

--------------------

When your development environment does not have the libelf library and you
compile an external module with the "M=xxx" compilation option, the
external module will not have ORC information and there will be no warning
or errors.

If a machine insmod this module and it's process is calling a function in
this module, typing kaptch will result in an error indicating that the
stack of this process is not trustworthy. The worst thing is that if this
module is a necessary and will be used for a long time, kpatch will be
difficult to type.

---------------------

Revert "kbuild: fix false positive warning/error about missing libelf"
This reverts commit ef7cfd00b2caf6edeb7f169682b64be2d0a798cf.

Fixes: ef7cfd00b2caf6edeb ("kbuild: fix false positive warning/error about missing libelf")
Signed-off-by: Zhao Xuedong <zhaoxuedong@meituan.com>