sched: skip update idle core in case of rq->nr_running is not zero
jingdong inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IAN7DY

--------------------------------

After enabling SMT expel, a new CPU state is introduced: the current is
idle, but there are offline tasks in the runqueue (rq) and nr_running is
not zero.

Before entering this state, the idle core status will be updated.
It is possible that the current core may be identified as an idle core
and marked as such on its local LLC (Last Level Cache).

In the fast path, when traversing idle cores, the kernel will deny
that this core is an idle core. There is a logical inconsistency here.

Fixes: fd5207be48fa ("sched: Implement the function of qos smt expeller")
Reviewed-by: zhaoxiaoqiang11 <zhaoxiaoqiang11@jd.com>
Signed-off-by: zhangwei123171 <zhangwei123171@jd.com>